API: Break phantom into submodules
