+++ src/xdesign/phantom/__init__.py
@@ -1,0 +1,10 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+"""Objects and methods for computing the quality of reconstructions.
+
+.. moduleauthor:: Daniel J Ching <carterbox@users.noreply.github.com>
+"""
+
+from xdesign.phantom.custom import *
+from xdesign.phantom.phantom import *
+from xdesign.phantom.standards import *
